% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qg_print_labels.R
\name{qg_print_labels}
\alias{qg_print_labels}
\title{qg_print_labels}
\usage{

  qg_print_labels(info = utils::read.csv(system.file("samples/invitro.csv",
  package = "qr4genebanks"), stringsAsFactors = FALSE),
  template = system.file("templates/label_seeds.yaml", package =
  "qr4genebanks"), to_file = "sample.pdf")
}
\arguments{
\item{info}{data.frame with content for barcode}

\item{template}{path to yaml file with configuration details}

\item{to_file}{path to output file}
}
\value{
nothing
}
\description{
Creates a PDF file with barcoded labels given a data.frame,
a template with formatting instructions and a name for the pdf file.
Currently, only QR code is supported.
}
\details{
All options for fine-tuning the label are contained within the template
file. The file uses a yaml format. Options can be adjusted for different
label sizes. The basic layout pattern consists of four blocks.

ID block

QR block

Info block

Optional info block
}
\examples{

if (interactive()) {
  qg_print_labels()
}
}
