% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qr4genebanks.R
\name{run_qr4genebanks}
\alias{run_qr4genebanks}
\title{run_qr4genebanks}
\usage{
run_qr4genebanks(port = 9866)
}
\arguments{
\item{port}{integer, default 9866}
}
\value{
NULL invisible
}
\description{
Runs a shiny app to print QR labels
}
\examples{

if(interactive()) {
   quagga::run_quagga()
}

}
\author{
Reinhard Simon
}
