% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_label.R
\name{print_label}
\alias{print_label}
\title{print_label}
\usage{
print_label(info, header = c("today", "Recursos Gen&#0233;ticos",
  "Banco in-vitro"), template = system.file("templates/label1.yaml", package =
  "quagga"), tofile = "sample.pdf")
}
\arguments{
\item{info}{data.frame with content for barcode}

\item{header}{vector of three strings which function as header}

\item{template}{path to yaml file with configuration details}

\item{tofile}{path to output file}
}
\value{
nothing
}
\description{
Creates a PDF file with barcoded labels given a data.frame,
a template with formatting instructions and a name for the pdf file.
Currently, only QR code is supported.
}
\details{
All options for fine-tuning the label are contained within the template
file. The file uses a yaml format. Options can be adjusted for different
label sizes. The basic layout pattern consists of four blocks.

ID block

QR block

Info block

Optional info block
}
\examples{

if (interactive()) {
info <- read.csv(
  system.file("/samples/invitro.csv", package = "quagga"),
  stringsAsFactors = FALSE)
print_label(info,
 header = c("Recursos Genéticos", "Banco in-vitro", "today"),
 template = "inst/templates/label1.yaml",
 tofile = "label1.pdf" )
}

}
